movie_match_bot/requirements.txt
aiogram==3.2.0
aiohttp==3.8.5
google-generativeai==0.3.0
python-dotenv==1.0.0